refactor(board): move from index to separate scenario
